Refactor module
This commit refactors the original model. It removes the
apache-specific code from the init module and moves it to
a separate apache namespace. It also adds the ability to
configure apache either as a normal alias via conf.d or as
a vhost.

Note that the vhost has *not* been tested.