PUPPETDB_SSL setting is now PUPPETDB_SSL_VERIFY
In the current master branch of puppetboard the PUPPETDB_SSL setting in
default_settings.py was renamed to PUPPETDB_SSL_VERIFY. This patch changes
this in the corresponding template but doesn't touch the $puppetdb_ssl
parameter of the puppetboard puppet class in init.pp for now.

Please consider changing the parameter name to $puppetdb_ssl_verify in
future releases.

Signed-off-by: Franz Pletz <fpletz@fnordicwalking.de>