Implement Python 3 support / Drop python 2
To support python3, we have to introduce a few new variables to support
this creation of a py3 virtualenv, and also to set wsgi python-home [1]
it. Just using activate_this.py didn't work.

In test, we also choose to use a separate virtualenv
`virtenv3-puppetboard` because `virtenv-puppetboard' will be created as
a py2 virtualenv by the other tests.

[1]: https://modwsgi.readthedocs.io/en/develop/user-guides/virtual-environments.html