Remove the $listen parameter
This parameter was used to adjust code directly in the puppetboard
checked-out files but the change has been broken by a [commit] in the
upstream project about 8 years ago.

This was not covered by unit tests so no need of tuning here.

Removing this should not harm users, but it still change the interface
of the module.

[commit]: https://github.com/voxpupuli/puppetboard/commit/cb64b738322710d3ee449c0bc9de7845665186ad