Do not manage $basedir when installing from packages
The $basedir directory is the root directory where pip/git fetch and
store all puppetboard data when installing with these tools, but is not
used when installing using system packages.

On FreeBSD, we default to install using packages, and the `/srv`
directory does not exist leading to catalog application falures by
default.

Since this directory is not used in such a case, do not manage it.